--- Optimal.Com.Web/Extensions/Extensions.cs
@@ -1,6 +1,0 @@
-﻿namespace Optimal.Com.Web.Extensions
-{
-    public class Extensions
-    {
-    }
-}
