+++ Optimal.Com.Web/Framework/Repository/IRepository.cs
@@ -1,0 +1,15 @@
+﻿using Optimal.Com.Web.Framework.Commons;
+
+namespace Optimal.Com.Web.Framework.Repository
+{
+    public interface IRepository<TEntity> where TEntity : BaseEntity
+    {
+        IQueryable<TEntity> Table { get; }
+        Task<IEnumerable<TEntity>> GetAllAsync();
+        Task<TEntity> GetByIdAsync(int id);
+        Task AddAsync(TEntity entity);
+        Task UpdateAsync(TEntity entity);
+        Task DeleteAsync(int id);
+        Task SaveAsync();
+    }
+}
